Split Api with Api & AlamofireApi
